Rename source ndarray_udfs directory
