+++ test/udfs/ndarray/test_crop.py
@@ -17,7 +17,7 @@ import unittest
 import numpy as np
 import pandas as pd
 
-from eva.udfs.ndarray_udfs.crop import Crop
+from eva.udfs.ndarray.crop import Crop
 
 
 class CropTests(unittest.TestCase):
